qr code updated

+++ frontend/src/components/Profile.js
@@ -12,11 +12,16 @@ import { useAuthStore } from "../redux/store1";
 import { updateUser, getProfile, deleteUser } from "../apicalls/helper";
 import QRCodeGenerator from "./QRCodeGenerator";
 import Modal from "react-modal";
+import { Button, Card, Col, Input, Row, Typography } from "antd";
+import QRCode from "qrcode";
+import jsQR from "jsqr";
 
 export default function Profile() {
   const navigate = useNavigate();
 
   const [deleteit, setDeleteit] = useState(false);
+  const [imageUrl, setImageUrl] = useState("");
+  const [text, setText] = useState("");
 
   // let apiData = {}
   const [apiData, setApiData] = useState({});
@@ -112,17 +117,26 @@ export default function Profile() {
   ///// QR code download
 
   // Create a ref to the QR code element
-  const qrCodeRef = useRef(null);
+  // const qrCodeRef = useRef(null);
 
-  // Function to handle the download button click
-  const handleDownloadClick = () => {
-    // Get the data URL of the QR code image
-    const qrCodeDataURL = qrCodeRef.current?.toDataURL("image/png");
+  // // Function to handle the download button click
+  // const handleDownloadClick = () => {
+  //   // Get the data URL of the QR code image
+  //   const qrCodeDataURL = qrCodeRef.current?.toDataURL("image/png");
 
-    const a = document.createElement("a");
-    a.href = qrCodeDataURL;
-    a.download = "qrcode.png";
-    a.click();
+  //   const a = document.createElement("a");
+  //   a.href = qrCodeDataURL;
+  //   a.download = "qrcode.png";
+  //   a.click();
+  // };
+
+  const generateQrCode = async () => {
+    try {
+      const response = await QRCode.toDataURL(apiData.studentId);
+      setImageUrl(response);
+    } catch (error) {
+      console.log(error);
+    }
   };
 
   return (
@@ -151,7 +165,7 @@ export default function Profile() {
             className={`${styles.glass} ${extend.glass}`}
             style={{ height: "98%" }}
           >
-            <div className="flex justify-center" >
+            <div className="flex justify-center">
               <div className="title flex flex-col items-center">
                 <h4 className="text-5xl font-bold"> Student Profile</h4>
                 <span className="py-4 text-xl w-2/3 text-center text-gray-500">
@@ -178,14 +192,23 @@ export default function Profile() {
               </div>
 
               <div style={{ float: "right", marginTop: "-200px" }}>
-                {/* <h1>Login Page</h1> */}
-                <QRCodeGenerator apiData={apiData} />
-                {/* <p>{apiData.name}</p> */}
-                {/* <p>{apiData.email}</p> */}
-                <p>{apiData.studentId}</p>
-                <button onClick={handleDownloadClick}>Download</button>
+                <Col className="col-span-8 lg:col-span-8 md:col-span-12 sm:col-span-24">
+                  {/* <Input placeholder="Enter text here" onChange={(e) => setText(e.target.value)} /> */}
+                  <Button
+                    className="mt-4 mb-10"
+                    variant="contained"
+                    color="primary-outlined"
+                    onClick={() => generateQrCode()}
+                  >
+                    Generate
+                  </Button>
+                  {imageUrl ? (
+                    <a href={imageUrl} download>
+                      <img className="w-24" src={imageUrl} alt="img" />
+                    </a>
+                  ) : null}
+                </Col>
               </div>
-
               <div className="textbox flex flex-col items-center gap-6">
                 <div className="name flex w-3/4 gap-10">
                   <input
@@ -252,9 +275,16 @@ export default function Profile() {
             </form>
             <button
               className={styles.btn}
-              style={{margin:'auto 12.5%',background:'rgba(155,12,20,0.5)'}}
+              style={{
+                margin: "auto 12.5%",
+                background: "rgba(155,12,20,0.5)",
+              }}
               onClick={() => {
                 // setDeleteit(true);
+                const confirmDelete=window.confirm("Are you sure you want to delete this?")
+                if(!confirmDelete){
+                  return
+                }
                 toast.promise(
                   deleteUser(apiData.id).then((res) => {
                     console.log(res);
